# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My docs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/transfer/active/update": {
            "post": {
                "tags": [
                    "Projects Transfer"
                ],
                "summary": "Обновить проекты и участников в состояние 'Активное'",
                "description": "Этот метод обновляет проекты и их участников, переводя проекты в состояние 'Активное' (state_id = 2). Участники, которые находятся в состоянии 'Подано' (state_id = 1), изменяются на состояние 'Отказано' (state_id = 4). Удаляются все заявки с состоянием 'Подано' (state_id = 1), которые не имеют альтернативных заявок на тот же проект с состоянием 'Отказано' (state_id = 4).",
                "operationId": "updateToActiveState",
                "responses": {
                    "200": {
                        "description": "Проекты и участники успешно обновлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated_projects": {
                                            "description": "Обновленные проекты",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID проекта",
                                                        "type": "integer"
                                                    },
                                                    "state_id": {
                                                        "description": "Состояние проекта",
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "description": "Название проекта",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "updated_count": {
                                            "description": "Количество обновленных проектов",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при обновлении проектов"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/transfer/approved/update": {
            "post": {
                "tags": [
                    "Projects Transfer"
                ],
                "summary": "Обновить активные проекты в состояние одобрены",
                "description": "Этот метод обновляет состояние проектов, клонирует их, сохраняет новые проекты в базу данных и изменяет состояние оригинальных проектов.",
                "operationId": "updateToApprovedProjects",
                "responses": {
                    "200": {
                        "description": "Проекты успешно обновлены, клонированы и сохранены"
                    },
                    "400": {
                        "description": "Ошибка при обновлении проектов"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/transfer/processing/update": {
            "post": {
                "tags": [
                    "Projects Transfer"
                ],
                "summary": "Обновить проекты в состояние 'Обработка'",
                "description": "Этот метод обновляет проекты, находящиеся в состоянии 'Набор' (state_id = 1), переводя их в состояние 'Обработка' (state_id = 5). Также обновляются соответствующие состояния кандидатов.",
                "operationId": "updateToProcessingState",
                "responses": {
                    "200": {
                        "description": "Проекты успешно обновлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated_projects": {
                                            "type": "array",
                                            "items": {
                                                "schema": "UpdateToProcessingStateController",
                                                "properties": {
                                                    "id": {
                                                        "description": "ID обновленного проекта",
                                                        "type": "integer"
                                                    },
                                                    "state_id": {
                                                        "description": "Новый ID состояния проекта",
                                                        "type": "integer"
                                                    },
                                                    "updated_at": {
                                                        "description": "Дата последнего обновления проекта",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "updated_count": {
                                            "description": "Количество обновленных проектов",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при загрузке или обработке данных"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
...
    },
...
}